% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{get_survey}
\alias{get_survey}
\title{Get a specific survey based on the survey name}
\usage{
get_survey(name)
}
\arguments{
\item{name}{The exact name of the survey, passed as a string. If the exact
name isn't supplied, an error is thrown with the list of all possible
surveys.}
}
\description{
Get a specific survey based on the survey name
}
