% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{set_key}
\alias{set_key}
\title{Set your API key to access the data from Qualtrics.}
\usage{
set_key(key, ...)
}
\arguments{
\item{key}{Your api key, passed as a string}

\item{...}{Additional arguments passed to \code{\link[qualtRics:qualtrics_api_credentials]{qualtRics::qualtrics_api_credentials()}}}
}
\description{
Saves your key in your .REnviron so it's accessible to later functions.
This only needs to be run once.
}
