% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpwdata-package.R
\docType{package}
\name{tpwdata-package}
\alias{tpwdata}
\alias{tpwdata-package}
\title{tpwdata: Grabs Survey data from the Turning Points for Women study}
\description{
This package downloads and prepares survey data from the Turning Points for Women study.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

Authors:
\itemize{
  \item Maria Schweer-Collins
}

}
\keyword{internal}
